test: update for playgrounds in js
