+++ src/editors/playgroundController.ts
@@ -5,7 +5,7 @@ import vm from 'vm';
 import * as os from 'os';
 
 import ActiveConnectionCodeLensProvider from './activeConnectionCodeLensProvider';
-import CodeActionProvider from './codeActionProvider';
+import PlaygroundSelectedCodeActionProvider from './playgroundSelectedCodeActionProvider';
 import ConnectionController, {
   DataServiceEventTypes,
 } from '../connectionController';
@@ -102,11 +102,11 @@ export default class PlaygroundController {
   _languageServerController: LanguageServerController;
   _selectedText?: string;
   _exportToLanguageCodeLensProvider: ExportToLanguageCodeLensProvider;
-  _codeActionProvider: CodeActionProvider;
+  _playgroundSelectedCodeActionProvider: PlaygroundSelectedCodeActionProvider;
+  _telemetryService: TelemetryService;
 
   _isPartialRun = false;
 
-  private _telemetryService: TelemetryService;
   private _activeConnectionCodeLensProvider: ActiveConnectionCodeLensProvider;
   private _outputChannel: OutputChannel;
   private _playgroundResultViewColumn?: vscode.ViewColumn;
@@ -125,7 +125,7 @@ export default class PlaygroundController {
     playgroundResultViewProvider: PlaygroundResultProvider,
     activeConnectionCodeLensProvider: ActiveConnectionCodeLensProvider,
     exportToLanguageCodeLensProvider: ExportToLanguageCodeLensProvider,
-    codeActionProvider: CodeActionProvider,
+    codeActionProvider: PlaygroundSelectedCodeActionProvider,
     explorerController: ExplorerController
   ) {
     this._connectionController = connectionController;
@@ -138,7 +138,7 @@ export default class PlaygroundController {
       vscode.window.createOutputChannel('Playground output');
     this._activeConnectionCodeLensProvider = activeConnectionCodeLensProvider;
     this._exportToLanguageCodeLensProvider = exportToLanguageCodeLensProvider;
-    this._codeActionProvider = codeActionProvider;
+    this._playgroundSelectedCodeActionProvider = codeActionProvider;
     this._explorerController = explorerController;
 
     this._connectionController.addEventListener(
@@ -157,23 +157,34 @@ export default class PlaygroundController {
       }
 
       if (isPlayground(editor?.document.uri)) {
-        void vscode.commands.executeCommand('setContext', 'mdb.showRunPlaygroundButton', true);
+        void vscode.commands.executeCommand(
+          'setContext',
+          'mdb.showRunPlaygroundButton',
+          true
+        );
       } else {
-        void vscode.commands.executeCommand('setContext', 'mdb.showRunPlaygroundButton', false);
+        void vscode.commands.executeCommand(
+          'setContext',
+          'mdb.showRunPlaygroundButton',
+          false
+        );
       }
 
       if (editor?.document.languageId !== 'Log') {
         this._activeTextEditor = editor;
+        this._activeConnectionCodeLensProvider.setActiveTextEditor(
+          this._activeTextEditor
+        );
+        this._playgroundSelectedCodeActionProvider.setActiveTextEditor(
+          this._activeTextEditor
+        );
         log.info('Active editor path', editor?.document.uri?.path);
       }
     };
 
     vscode.workspace.textDocuments.forEach((document) => {
       if (isPlayground(document.uri)) {
-        void vscode.languages.setTextDocumentLanguage(
-          document,
-          'javascript'
-        );
+        void vscode.languages.setTextDocumentLanguage(document, 'javascript');
       }
     });
 
@@ -182,16 +193,20 @@ export default class PlaygroundController {
 
     vscode.workspace.onDidOpenTextDocument(async (document) => {
       if (isPlayground(document.uri)) {
-        await vscode.languages.setTextDocumentLanguage(
-          document,
-          'javascript'
-        );
+        this._telemetryService.trackPlaygroundLoaded();
+        await vscode.languages.setTextDocumentLanguage(document, 'javascript');
+      }
+    });
+
+    vscode.workspace.onDidSaveTextDocument((document) => {
+      if (isPlayground(document.uri)) {
+        this._telemetryService.trackPlaygroundSaved();
       }
     });
 
     vscode.window.onDidChangeTextEditorSelection(
       async (changeEvent: vscode.TextEditorSelectionChangeEvent) => {
-        if (isPlayground(changeEvent?.textEditor?.document?.uri)) {
+        if (!isPlayground(changeEvent?.textEditor?.document?.uri)) {
           return;
         }
 
@@ -216,7 +231,7 @@ export default class PlaygroundController {
             selection: sortedSelections[0],
           });
 
-        this._codeActionProvider.refresh({
+        this._playgroundSelectedCodeActionProvider.refresh({
           selection: sortedSelections[0],
           mode,
         });
@@ -265,12 +280,20 @@ export default class PlaygroundController {
       const workspaceFolder = vscode.workspace.workspaceFolders?.[0];
       const filePath = workspaceFolder?.uri.fsPath || os.tmpdir();
 
-      const numberUntitledPlaygrounds = vscode.workspace.textDocuments.filter((doc) => isPlayground(doc.uri)).length;
-      const fileName = path.join(filePath, `playground-${numberUntitledPlaygrounds + 1}.mongodb.js`);
+      const numberUntitledPlaygrounds = vscode.workspace.textDocuments.filter(
+        (doc) => isPlayground(doc.uri)
+      ).length;
+      const fileName = path.join(
+        filePath,
+        `playground-${numberUntitledPlaygrounds + 1}.mongodb.js`
+      );
 
       // Does not create a physical file, it only creates a URI from specified component parts.
       // An untitled file URI: untitled:/extensionPath/playground-1.mongodb.js
-      const documentUri = vscode.Uri.from({ path: fileName, scheme: 'untitled' });
+      const documentUri = vscode.Uri.from({
+        path: fileName,
+        scheme: 'untitled',
+      });
 
       // Fill in initial content.
       const edit = new vscode.WorkspaceEdit();
@@ -578,7 +601,10 @@ export default class PlaygroundController {
   }
 
   runAllPlaygroundBlocks(): Promise<boolean> {
-    if (!this._activeTextEditor || !isPlayground(this._activeTextEditor.document.uri)) {
+    if (
+      !this._activeTextEditor ||
+      !isPlayground(this._activeTextEditor.document.uri)
+    ) {
       void vscode.window.showErrorMessage(
         "Please open a '.mongodb' playground file before running it."
       );
@@ -593,7 +619,10 @@ export default class PlaygroundController {
   }
 
   runAllOrSelectedPlaygroundBlocks(): Promise<boolean> {
-    if (!this._activeTextEditor || !isPlayground(this._activeTextEditor.document.uri)) {
+    if (
+      !this._activeTextEditor ||
+      !isPlayground(this._activeTextEditor.document.uri)
+    ) {
       void vscode.window.showErrorMessage(
         "Please open a '.mongodb' playground file before running it."
       );
@@ -647,9 +676,9 @@ export default class PlaygroundController {
       ...this._exportToLanguageCodeLensProvider._exportToLanguageAddons,
       textFromEditor: this._getAllText(),
       selectedText: this._selectedText,
-      selection: this._codeActionProvider.selection,
+      selection: this._playgroundSelectedCodeActionProvider.selection,
       language,
-      mode: this._codeActionProvider.mode,
+      mode: this._playgroundSelectedCodeActionProvider.mode,
     });
 
     return this._transpile();
@@ -700,9 +729,15 @@ export default class PlaygroundController {
         },
       };
 
-      if (this._codeActionProvider.mode === ExportToLanguageMode.AGGREGATION) {
+      if (
+        this._playgroundSelectedCodeActionProvider.mode ===
+        ExportToLanguageMode.AGGREGATION
+      ) {
         toCompile.aggregation = selectedText;
-      } else if (this._codeActionProvider.mode === ExportToLanguageMode.QUERY) {
+      } else if (
+        this._playgroundSelectedCodeActionProvider.mode ===
+        ExportToLanguageMode.QUERY
+      ) {
         toCompile.filter = selectedText;
       }
 
@@ -758,7 +793,10 @@ export default class PlaygroundController {
       log.info(`Export to ${language} language result`, this._playgroundResult);
 
       /* eslint-disable camelcase */
-      if (this._codeActionProvider.mode === ExportToLanguageMode.AGGREGATION) {
+      if (
+        this._playgroundSelectedCodeActionProvider.mode ===
+        ExportToLanguageMode.AGGREGATION
+      ) {
         const aggExportedProps = {
           language,
           num_stages: selectedText
@@ -770,7 +808,10 @@ export default class PlaygroundController {
         };
 
         this._telemetryService.trackAggregationExported(aggExportedProps);
-      } else if (this._codeActionProvider.mode === ExportToLanguageMode.QUERY) {
+      } else if (
+        this._playgroundSelectedCodeActionProvider.mode ===
+        ExportToLanguageMode.QUERY
+      ) {
         const queryExportedProps = {
           language,
           with_import_statements: importStatements,
