+++ src/telemetry/telemetryService.ts
@@ -104,22 +104,6 @@ export default class TelemetryService {
     this._segmentUserId = userId;
     this._segmentAnonymousId = anonymousId;
     this._segmentKey = this._readSegmentKey();
-
-    vscode.workspace.onDidOpenTextDocument((document) => {
-      if (
-        document &&
-        document.languageId === 'mongodb' &&
-        document.uri.scheme === 'file'
-      ) {
-        this.trackPlaygroundLoaded();
-      }
-    });
-
-    vscode.workspace.onDidSaveTextDocument((document) => {
-      if (document && document.languageId === 'mongodb') {
-        this.trackPlaygroundSaved();
-      }
-    });
   }
 
   private _readSegmentKey(): string | undefined {
