+++ src/test/suite/telemetry/telemetryService.test.ts
@@ -60,12 +60,12 @@ suite('Telemetry Controller Test Suite', () => {
       mockTrackCommandRun
     );
     sinon.replace(
-      mdbTestExtension.testExtensionController._telemetryService,
+      mdbTestExtension.testExtensionController._playgroundController._telemetryService,
       'trackPlaygroundCodeExecuted',
       mockTrackPlaygroundCodeExecuted
     );
     sinon.replace(
-      mdbTestExtension.testExtensionController._telemetryService,
+      mdbTestExtension.testExtensionController._playgroundController._telemetryService,
       'trackPlaygroundLoaded',
       mockTrackPlaygroundLoadedMethod
     );
@@ -318,7 +318,7 @@ suite('Telemetry Controller Test Suite', () => {
     const mode = ExportToLanguageMode.QUERY;
     const language = 'python';
 
-    mdbTestExtension.testExtensionController._playgroundController._codeActionProvider.mode =
+    mdbTestExtension.testExtensionController._playgroundController._playgroundSelectedCodeActionProvider.mode =
       mode;
     mdbTestExtension.testExtensionController._playgroundController._exportToLanguageCodeLensProvider._exportToLanguageAddons =
       {
@@ -361,7 +361,7 @@ suite('Telemetry Controller Test Suite', () => {
     const mode = ExportToLanguageMode.AGGREGATION;
     const language = 'java';
 
-    mdbTestExtension.testExtensionController._playgroundController._codeActionProvider.mode =
+    mdbTestExtension.testExtensionController._playgroundController._playgroundSelectedCodeActionProvider.mode =
       mode;
     mdbTestExtension.testExtensionController._playgroundController._exportToLanguageCodeLensProvider._exportToLanguageAddons =
       {
@@ -384,47 +384,6 @@ suite('Telemetry Controller Test Suite', () => {
       with_import_statements: false,
       with_builders: false,
       with_driver_syntax: false,
-    });
-  });
-
-  suite('with active connection', function () {
-    this.timeout(5000);
-
-    let dataServ;
-    beforeEach(async () => {
-      try {
-        dataServ = await connect({ connectionString: TEST_DATABASE_URI });
-      } catch (error) {
-        expect(error).to.be.undefined;
-      }
-    });
-
-    afterEach(async () => {
-      sinon.restore();
-      await dataServ.disconnect();
-    });
-
-    test('track new connection event fetches the connection instance information', async () => {
-      sinon.replace(testTelemetryService, 'track', mockTrack);
-      sinon.replace(
-        testTelemetryService,
-        '_isTelemetryFeatureEnabled',
-        () => true
-      );
-      await mdbTestExtension.testExtensionController._telemetryService.trackNewConnection(
-        dataServ,
-        ConnectionTypes.CONNECTION_STRING
-      );
-
-      expect(mockTrack.firstCall.args[0]).to.equal('New Connection');
-      const instanceTelemetry: NewConnectionTelemetryEventProperties =
-        mockTrack.firstCall.args[1];
-      expect(instanceTelemetry.is_localhost).to.equal(true);
-      expect(instanceTelemetry.is_atlas).to.equal(false);
-      expect(instanceTelemetry.is_used_connect_screen).to.equal(false);
-      expect(instanceTelemetry.is_used_command_palette).to.equal(true);
-      expect(instanceTelemetry.is_used_saved_connection).to.equal(false);
-      expect(instanceTelemetry.is_genuine).to.equal(true);
     });
   });
